Type and fix tests
